==> 06_inStrain_RUG4941/01_stb_file.ipynb <==
import os
from Bio import SeqIO
import glob
import pandas as pd

os.chdir("/fs/scratch/PAS0439/Ming/databases/RUG4941_rumen_mags_hungate_high_quality/drep_for_instrain/dereplicated_genomes")

genomes = glob.glob("*")
omit = {'scaffold_10121|kraken:taxid|860848478':0, 'scaffold_2871|kraken:taxid|3782448846':0,  'nrc_11885|kraken:taxid|2432288124':0}

with open("../../../../virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/genomes.stb", "w") as outfile:
    for f in genomes:
        records = SeqIO.parse(f, "fasta")
        for record in records:
            if record.id not in omit.keys():
                outfile.write(record.id + "\t" + f + '\n')
            else:
                omit[record.id] += 1
                if omit[record.id] > 1:
                    continue
                else:
                    outfile.write(record.id + "\t" + f + '\n')
