==> 06_inStrain_RUG4941/02_subset_for_instrain_compare.ipynb <==
import os
import pandas as pd
import glob

os.chdir("/fs/scratch/PAS0439/Ming/databases/RUG4941_rumen_mags_hungate_high_quality/drep_for_instrain/dereplicated_genomes")

drep =  [f.split(".fasta")[0] for f in glob.glob("*.fasta")]

taxa = pd.read_csv("/fs/ess/PAS0439/MING/databases/hungate_rug_rumen_mags_taxa.csv")

drep_taxa = taxa.query('user_genome in @drep')

prophage_mags = pd.read_csv("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/prophage_host_summary.csv")

prophage_species = list(prophage_mags.query('user_genome in @drep').query('prophage == prophage').Species)

prophage_species_multiple_mags = drep_taxa.query('Species in @prophage_species').groupby("Species").user_genome.count().reset_index().query('user_genome >1').Species

genome_for_instrain_compare = drep_taxa.query('Species in @prophage_species_multiple_mags')

genome_for_instrain_compare[genome_for_instrain_compare.user_genome.str.contains("RUG")].query('Genera == "Prevotella"')
