==> 06_inStrain_RUG4941/03_microdiversity_prevelance.ipynb <==
import os
import pandas as pd
import glob

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/instrain_profiles/")

instrains = glob.glob("**/*genome_info.tsv*", recursive = True)

df_list = []
for f in instrains:
    id = f.split("/")[0].split(".profile")[0]
    df = pd.read_csv(f,  sep = "\t")
    df.genome = df.genome.apply(lambda x: x.split(".fasta")[0])
    df = df[["genome", "coverage", "breadth", "nucl_diversity"]]
    df.loc[:,"id"] = id
    df_list.append(df)

instrain_out = pd.concat(df_list).reset_index(drop = True)

genome_taxa = pd.read_csv("../genome_info.csv")[["user_genome", "Genera", "Species"]]

instrain_taxa = pd.merge(genome_taxa, instrain_out, left_on = "user_genome", right_on = "genome").drop("user_genome", axis = 1)

species_pervelance = pd.read_csv("../../R_project/microbe_species_prevelance_df.csv")[["Domain","Species", "prevelance"]]

instrain_taxa_prevelance = pd.merge(instrain_taxa, species_pervelance, on = "Species")

metadata = pd.read_csv("../../R_project/metadata_ecology.csv")

microdiversity = pd.merge(instrain_taxa_prevelance, metadata, on = "id").query("coverage >5")

microdiversity.to_csv("../../R_project/microdiversity_stewart_rug4941.csv", index = None) 

microdiversity.to_csv("/fs/ess/PAS0439/MING/virome_ecology/results/R_project/microdiversity_stewart_rug4941.csv", index = None)
