==> 06_inStrain_RUG4941/04_parse_dbcan_rug4941.ipynb <==
import os
import pandas as pd

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/")

df = pd.read_csv("repre_genomes_instrain.genes.faa_vs_dbCAN_v11.csv")

pd.set_option('display.max_columns', None)

cazy = df[["column0", "column2", "column3", "column5", "column6", "column15","column16","column17","column18"]]
cazy.columns = ["CAZyme", "qlen", "gene", "slen", "E_value", "qstart", "qend", "sstart", "send"]

for index, row in cazy.iterrows():
    cazy.loc[index, "hmm_coverage"] = (cazy.loc[index, "qend"] - cazy.loc[index, "qstart"]) / cazy.loc[index, "qlen"]

cazy_ready = cazy[["CAZyme", "gene", "E_value", "hmm_coverage"]]

cazy_ready.loc[:,"CAZyme"] = list(cazy_ready.CAZyme.apply(lambda x: x.split(".hmm")[0]))

cazy_ready_filtered = cazy_ready.query('E_value == E_value')

cazy_hits = cazy_ready_filtered.loc[cazy_ready_filtered.groupby("gene")["E_value"].idxmin()]

gh = cazy_hits[cazy_ready_filtered.CAZyme.str.contains("GH")]

gh_ready = gh.query('E_value < 1e-15').query('hmm_coverage > 0.35').reset_index(drop = True)

gh_ready

gh_ready.to_csv("dbcan_out_strigent_RUG4941.csv", index = None)

gh_ready[["gene", "CAZyme"]].rename(columns = {"CAZyme":"anno"}).to_csv("geneAnnotations_CAZyme.csv", index = None)
