==> 06_inStrain_RUG4941/08_prophage_gene_under_selection.ipynb <==
import os
import pandas as pd
import glob
from Bio import SeqIO

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/prophage_carrying_genome_annotate/dram_anno/")

prophage_co_exist_microdiversity = pd.read_csv("/fs/ess/PAS0439/MING/virome_ecology/results/R_project/prophage_co_exist_microdiversity_rug4941_tmp.csv")
prophage_microdiversity = pd.read_csv("/fs/ess/PAS0439/MING/virome_ecology/results/R_project/prophage_microdiversity_rug4941.csv_tmp")
instrain_genome_taxa = pd.read_csv("../../genome_info.csv")

pd.set_option('display.max_columns', None)

gene_under_selection = set(prophage_microdiversity.query('prophage_gene == "Yes"').query('SNV_count >= 10').query('pNpS_variants > 1').gene)
n_gene_under_selection = len(gene_under_selection)

records = SeqIO.parse("../../repre_genomes_instrain.genes.faa", "fasta")
with open("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/prophage_carrying_genome_annotate/gene_under_selection.faa", "w") as outfile:
    for record in records:
        if record.id in gene_under_selection:
            SeqIO.write(record, outfile, "fasta")

n_gene_qualified = len(set(prophage_microdiversity.query('prophage_gene == "Yes"').query('SNV_count >= 10').gene))

f"Out of {n_gene_qualified} genes, {n_gene_under_selection} are under postive selection"

*For species which both prophage-carrying and non-carrying population co-exist, whether prophage gene help specification*

#prophage_co_exist_microdiversity.query('prophage_gene == "Yes"').query('SNV_count > 10').query('pNpS_variants > 1').groupby("gene").id.nunique("id").reset_index()

*For each prophage-carrying population among the co-exist species, how many samples they are presented*

prophage_co_exist_microdiversity.query('prophage_gene == "Yes"')[["genome", "id"]].drop_duplicates().groupby("genome").nunique("id")

instrain_genome_taxa.query('user_genome == "GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic"')

*Prophage-carrying genome (MAG), GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic.fasta  
co-exist with prophage-non-carrying MAGs (of the same species: Prevotella sp900317685) in 46 samples*

*For GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic.fasta, any gene under selection?*

prophage_co_exist_microdiversity.query('genome == "GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic"').query('prophage_gene == "Yes"').query('SNV_count >= 10').groupby("gene").pNpS_variants.describe() 

prophage_co_exist_microdiversity_for_plot = prophage_co_exist_microdiversity.query('genome == "GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic"').query('prophage_gene == "Yes"').query('SNV_count >= 10')

*Save MAG GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic.fasta prophage pn/ps to plot*

prophage_co_exist_microdiversity_for_plot[['scaffold', 'gene', 'start', 'end', 'direction', 'pNpS_variants']].fillna(0).to_csv("/fs/ess/PAS0439/MING/virome_ecology/results/R_project/co_existing_prophage_ex_pnps.csv", index = None)

*MAG GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic.fasta prophage for gggenome*

anno = pd.read_csv("../dram_anno_genomes/GCA_902777695.1_Rumen_uncultured_genome_RUG12296_genomic.fasta/annotations.tsv", sep = "\t")

template_prophage_anno = anno.query('scaffold == "CACYVA010000059.1"').query(' 1<= gene_position <= 17')[["fasta", "scaffold", "gene_position", "start_position", "end_position", "strandedness", "vogdb_hit"]]
template_prophage_anno.to_csv("/fs/ess/PAS0439/MING/virome_ecology/results/R_project/co_existing_prophage_ex_for_gggenome.csv", index = None)

template_prophage_anno

*How many samples have species which both prophage-carrying and non-carrying population co-exist?  
For each sample, how many co-existing species?*

prophage_co_exist_present = prophage_co_exist_microdiversity[["Species", "id"]].drop_duplicates().groupby("id").nunique("Species")

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/instrain_profiles/")
df_list = []
for f in glob.glob("*"):
    i = f.split(".profile")[0]
    path = f"{i}.profile/output/{i}.profile_genome_info.tsv"
    if os.path.exists(path):
        df = pd.read_csv(f"{i}.profile/output/{i}.profile_genome_info.tsv", sep = "\t").query('breadth_minCov >= 0.5')   
    else:
        path = f"{i}.profile/output/{i}.profile_genome_info.tsv.gz"
        df = pd.read_csv(f"{path}", sep = "\t").query('breadth_minCov >= 0.5')
    df.loc[:, "id"] = i
    df_list.append(df)
instrain_genome = pd.concat(df_list)
instrain_genome.genome = [f.split(".fasta")[0] for f in list(instrain_genome.genome)]

instrain_genome_taxa = pd.read_csv("/fs/ess/PAS0439/MING/databases/hungate_rug_rumen_mags_taxa.csv")
instrain_genome_final = pd.merge(instrain_genome, instrain_genome_taxa[["user_genome", "Genera", "Species"]], left_on = "genome", right_on = "user_genome")

instrain_genome_present = instrain_genome_final[["Species", "id"]].drop_duplicates().groupby("id").nunique("Species")

prophage_co_exist_present.rename(columns = {"Species": "prophage_carrying"}, inplace = True)

instrain_genome_present.rename(columns = {"Species": "total"}, inplace = True)

prophage_co_exist_present_percent = pd.merge(prophage_co_exist_present, instrain_genome_present, on = "id")

for index, row in prophage_co_exist_present_percent.iterrows():
    prophage_co_exist_present_percent.loc[index,"percent"] = prophage_co_exist_present_percent.loc[index,"prophage_carrying"]/prophage_co_exist_present_percent.loc[index,"total"]*100

prophage_co_exist_present_percent

prophage_co_exist_present_percent.to_csv("/fs/ess/PAS0439/MING/virome_ecology/results/R_project/prophage_co_exist_present_percent.csv", index = None)

co_exist_prophage_gene_under_selection = set(prophage_co_exist_microdiversity.query('prophage_gene == "Yes"').query("divergent_site_count >=  10").query('pNpS_variants > 1').gene)
n_co_exist_prophage_gene_under_selection = len(co_exist_prophage_gene_under_selection)

co_exist_prophage_gene_qualified = set(prophage_co_exist_microdiversity.query('prophage_gene == "Yes"').query("divergent_site_count >=  10").gene)
n_co_exist_prophage_gene_qualified = len(co_exist_prophage_gene_qualified)

f"Out of {n_co_exist_prophage_gene_qualified} genes, {n_co_exist_prophage_gene_under_selection} are under postive selection"

records = SeqIO.parse("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/repre_genomes_instrain.genes.faa", "fasta")
with open("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/08_inStrain_rug4941/prophage_carrying_genome_annotate/coexist_prophage_gene_under_selection.faa", "w") as outfile:
    for record in records:
        if record.id in co_exist_prophage_gene_under_selection:
            SeqIO.write(record, outfile, "fasta")
